requirements.txt
torch
torchvision
torchaudio
streamlit
pillow
fastapi
uvicorn
